+++ R/dict_recode.R
@@ -2,13 +2,20 @@
 #' @inheritParams recode_na_as 
 #' @param patterns A vector containing patterns to use for pattern_type
 #' @param values A vector containing values to match to the patterns vector
-#' @return A `data.frame` object with replacements as required. 
+#' @param use_func Function to use for the recoding. One of the various 
+#' `recode_*` functions in package `mde`. 
+#' @return A `data.frame` object with replacements as required.
 #' @examples 
 #' head(dict_recode(airquality, pattern_type="starts_with",
-#' patterns = c("Solar", "Ozone"), values = c(420, 42)))
+#' patterns = c("Solar", "Ozone"), values = c(190, 41),
+#' use_func="recode_as_na"))
+#'  head(dict_recode(airquality, pattern_type="starts_with",
+#' patterns = c("Solar", "Ozone"), values = c(42, 420),
+#' use_func="recode_na_as"))
 #' @export
 
 dict_recode <- function(df, 
+                        use_func= "recode_na_as",
                         pattern_type="starts_with", 
                         patterns,
                         values){
@@ -18,6 +25,7 @@ dict_recode <- function(df,
 #' @export
 
 dict_recode.data.frame <- function(df, 
+                                   use_func = "recode_na_as",
                                    pattern_type="starts_with",
                                    patterns,
                                    values){
@@ -33,10 +41,9 @@ cur_col <-recode_selectors(names(df),
                    pattern=pat, 
                    column_check=FALSE)
 
-recode_na_as(df,
-             pattern_type = pattern_type,
-             pattern = pat, 
-             value= val)[cur_col]}, patterns, values)
+do.call(use_func, list(df=df, pattern_type=pattern_type,
+                             pattern=pat,value=val))[cur_col]}, 
+patterns, values)
 
 
 # Convert to a data.frame object from a list of vectors